Fix: #69 - Searchbar 순수하게 변경
- 원래 인풋, 드랍다운 까지만 컴포넌트내에서 해결하도록 수정
- 드랍다운 아이템에 대한 온클릭 핸들러를 프롭스로 받도록 수정
